\name{C5ClassifierDefinition}
\alias{C5ClassifierDefinition}
\title{Read the classifier definition from the file.}
\usage{
C5ClassifierDefinition(ch5file, channel_region, ...)
}
\arguments{
  \item{ch5file}{\code{\link{CellH5}} instance}
  \item{channel_region}{A channel-region pair e.g. \code{primary__primary}.}
  }
\author{Rudolf Hoefler <rudolf.hoefler@imba.oeaw.ac.at>}
