\name{C5Timelapse}
\alias{C5Timelapse}
\title{Read the timelapse table for a given positions}
\description{
Read the timelapse table as data.frame for a given position
}
\usage{
timelapse <- C5Timelapse(position)
}
\arguments{
  \item{position}{\code{item of \link{C5Positions}}}
}
\author{Rudolf Hoefler <rudolf.hoefler@imba.oeaw.ac.at>}