\name{CellH5}
\alias{CellH5}
\title{S4 class for easy access of ch5 files.}
\description{
This package provides a S4 class various convenience functions to read the data 
from the files.
}
\usage{
c5f <- CellH5(filepath) 
}
\note{

}
\arguments{
  \item{file}{ absolute or relative path to and ch5 file (or hdf5 file)}}
\author{Rudolf Hoefler <rudolf.hoefler@imba.oeaw.ac.at>}
\seealso{
  \code{\link{C5Close}},
  \code{\link{C5FileInfo}},
  \code{\link{C5Plates}},
  \code{\link{C5Positions}},}
